requirements.txt
numpy==1.24.3
Pillow==9.4.0
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.180
